Restructure CAN code into comms_CAN
