+++ speeduino/comms_secondary.h
@@ -29,9 +29,6 @@
 #endif
 
 void secondserial_Command(void);//This is the heart of the Command Line Interpreter.  All that needed to be done was to make it human readable.
-void can_Command(void);
 void sendCancommand(uint8_t cmdtype , uint16_t canadddress, uint8_t candata1, uint8_t candata2, uint16_t sourcecanAddress);
-void obd_response(uint8_t therequestedPID , uint8_t therequestedPIDlow, uint8_t therequestedPIDhigh);
-void readAuxCanBus();
 
 #endif // COMMS_SECONDARY_H
